SwiftUI Alarm App Splash Screen
This GitHub repository contains the SwiftUI code for a stylish and animated splash screen for an Alarm App. The splash screen features custom fonts (Alcatraz and TrackdriftRegular), a captivating welcome image, and animated gradient-filled circles. The code is designed to create an engaging and visually appealing introduction to the Alarm App, complete with animations and transitions. Explore the code to see how SwiftUI can be used to create an attractive splash screen for your iOS app.